Added Loading Level Prefab
For all new levels to be strung together. Place the Level Loader prefab in them at the end of the level. Align the bottom of the collider with the bottom of the floor. When the momento is dropped it will hit the collider and cause a new level to be loaded. My hotfix was removed from memento.cs